Hooked up socket activation

+++ Source/Catapult.AvaloniaApp/SocketActivation.cs
@@ -1,21 +1,26 @@
 using System;
 using System.IO;
 using System.Net.Sockets;
+using System.Runtime.InteropServices;
 using System.Text;
+using System.Threading;
 
 namespace Catapult.AvaloniaApp
 {
     public class SocketActivation
     {
-        static string SocketPath = "/tmp/catapultsocket";
+        static string SocketPath = Path.Combine(Path.GetTempPath(), "catapultsocket");
+        const string SocketToken = "CATAPULT";
 
-        static bool TryConnect()
+        public static Action Activator { get; internal set; }
+
+        public static bool TryConnect()
         {
             try
             {
                 var listener = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
                 listener.Connect(new UnixDomainSocketEndPoint("/tmp/testsocket"));
-                listener.Send(Encoding.UTF8.GetBytes("<EOF>"));
+                listener.Send(Encoding.UTF8.GetBytes(SocketToken));
                 listener.Disconnect(false);
                 return true;
             }
@@ -26,37 +31,49 @@ namespace Catapult.AvaloniaApp
             }
         }
 
-        static void CreateServer()
+        public static void CreateServer()
         {
-            File.Delete(SocketPath);
-
-            var listener = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
-            listener.Bind(new UnixDomainSocketEndPoint(SocketPath));
-            listener.Listen(1);
-
-            byte[] bytes = new Byte[1024];
-
-            // Start listening for connections.
-            while (true)
+            if (!RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
             {
-                Socket handler = listener.Accept();
-                string data = null;
+                return;
+            }
 
-                // An incoming connection needs to be processed.
+            var thread = new Thread(() =>
+            {
+                File.Delete(SocketPath);
+
+                var listener = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
+                listener.Bind(new UnixDomainSocketEndPoint(SocketPath));
+                listener.Listen(1);
+
+                byte[] bytes = new Byte[1024];
+
+                // Start listening for connections.
                 while (true)
                 {
-                    int bytesRec = handler.Receive(bytes);
-                    data += Encoding.UTF8.GetString(bytes, 0, bytesRec);
+                    Socket handler = listener.Accept();
+                    string data = null;
 
-                    if (data.IndexOf("<EOF>") > -1)
+                    // An incoming connection needs to be processed.
+                    while (true)
                     {
-                        break;
-                    }
-                }
+                        int bytesRec = handler.Receive(bytes);
+                        data += Encoding.UTF8.GetString(bytes, 0, bytesRec);
 
-                handler.Shutdown(SocketShutdown.Both);
-                handler.Close();
-            }
+                        if (data.IndexOf(SocketToken) > -1)
+                        {
+                            break;
+                        }
+
+                        Activator?.Invoke();
+                    }
+
+                    handler.Shutdown(SocketShutdown.Both);
+                    handler.Close();
+                }
+            }) { IsBackground = true };
+
+            thread.Start();
         }
     }
 }
